requirements.txt
requests
pytest
python-dotenv
allure-pytest
pydantic
loguru
